+++ tests/testthat/test-gen_results.R
@@ -4,13 +4,13 @@ test_that("gen_results works", {
 
   gene_data <- HPOExplorer::load_phenotype_to_genes()
   ctd <- load_example_ctd()
-  list_names <- unique(gene_data$hpo_name)[seq(3)]
+  list_names <- unique(gene_data$hpo_id)[seq(3)]
   all_results <- gen_results(ctd = ctd,
                              gene_data = gene_data,
                              list_names = list_names,
                              reps = 10)
   testthat::expect_true(methods::is(all_results,"data.table"))
-  testthat::expect_gte(sum(list_names %in% unique(all_results$hpo_name)),
+  testthat::expect_gte(sum(list_names %in% unique(all_results$hpo_id)),
                        length(list_names)-1)
   testthat::expect_gte(nrow(all_results[q<=0.05,]),10)
 })