+++ tests/testthat/test-load_hpo_graph.R
@@ -1,0 +1,5 @@
+test_that("load_hpo_graph works", {
+
+  g <- load_hpo_graph()
+  testthat::expect_true(methods::is(g,"igraph"))
+})